% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{abundanceImage}
\alias{abundanceImage}
\title{Construct an image from a matrix of spectra using selected bands}
\usage{
abundanceImage(M = NULL, nrow = 0, ncol = 0, band = 1:ncol(M), ...)
}
\arguments{
\item{M}{Matrix containing spectra in rows.
The matrix should have dimensions no. of samples x
no. of wavelengths.}

\item{nrow}{Integer. The number of rows in the final image.}

\item{ncol}{Integer. The number of columns in the final image.}

\item{band}{Integer.  A vector of integers giving the bands to be
displayed in the image.  The intensity of the specified bands are
summed to give a value which will be the intensity in the image.
Defaults to all bands.  One can also use any expression that
evaluates to a vector of \code{TRUE/FALSE} for subsetting.}

\item{...}{Arguments to be passed downstream.}
}
\value{
A matrix suitable for display with \code{image}.
}
\description{
Convert a set of spectra into a form suitable for
making a false color image.  The spectra are first
subset according to the requested bands (wavelengths/frequencies).
Then, the spectra are
stacked into columns (as R is a column-major language)
to create the final image, which is a matrix ready
to be passed to \code{image}.
}

