% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.hclust.R
\name{as.hclust}
\alias{as.hclust}
\title{Convert exact linkage results to \code{hclust} objects}
\usage{
as.hclust(x, labels = 1:(nrow(x) + 1))
}
\arguments{
\item{x}{a matrix with each pair of leaves/nodes (first two columns) and
the distance between them (third column)}

\item{labels}{a unique label that identifies each leaf}
}
\value{
A \code{hclust} object
}
\description{
Converts the output of the \code{elink} function to a \code{hclust} object.
}
\seealso{
\code{\link{elink}}
}
