% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.dist.R
\name{as_dist}
\alias{as_dist}
\title{Calculate distance matrix from class assignments}
\usage{
as_dist(x, labels)
}
\arguments{
\item{x}{a matrix of class assignments, as produced by \code{DPM_sampler()}}

\item{labels}{a label for each individual}
}
\value{
an object of class \code{dist}
}
\description{
Takes the class assignments for each individual generated by the DPM sampler
and computes a distance matrix between individuals.
}
\details{
Distance is calculated as the proportion of assignments in which each pair of
individuals gets assigned to the same class.  Note that \code{elink()} is the
preferred method to do this.
}
\seealso{
\link{DPM_sampler}, \link{elink}
}
