% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.hclust.R
\name{as_hclust}
\alias{as_hclust}
\title{Hierarchical clustering of classification results}
\usage{
as_hclust(x, N, labels = 1:N)
}
\arguments{
\item{x}{a matrix with three columns, in which the first two are integers
identifying the individuals in each pair and the third is the distance
between them.  Such a matrix can easily be obtained by \code{cbind}ing the
two objects output by \code{elink()}.}

\item{N}{the number of individuals}

\item{labels}{a vector of labels to be assigned to each individual.  Defaults
to 1:N}
}
\value{
a \code{hclust} object
}
\description{
Performs a hierarchical clustering of individuals from the distances between
pairs.
}
\seealso{
\link{elink}
}
