% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{categorical_model}
\alias{categorical_model}
\title{Constructor for Categorical models}
\usage{
categorical_model(n_components, type = "bernoulli", max_val = NULL, ...)
}
\description{
Constructor for Categorical models
}
