% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{cos.dist}
\alias{cos.dist}
\title{cosine distance}
\usage{
\method{cos}{dist}(x, ...)
}
\arguments{
\item{x}{matrix (or object that can be coerced into matrix with \code{as.matrix (x)}

cosine distance is computed for the rows of \code{x}}

\item{...}{ignored}
}
\value{
dist object: lower triangular matrix with pairwise distances
}
\description{
computes the cosine distance matrix between the rows of \code{x}.
}
\details{
Cosine distance between two row vectors a and b is defined as

\deqn{cos \alpha = \frac{a \times b^T}{\left||a\right|| \cdot \left||b\right||}}{cos alpha = (a x t (b)) / (||a|| * ||b||)}

with the vector norms
\deqn{\left||x\right|| = \sqrt{\sum x_i^2}}{||x|| = sqrt (sum (x[i]^2))}
}
\author{
Claudia Beleites
}
\seealso{
See e.g.

Varmuza, K. and Filzmoser, P.: Introduction to multivariate statistical analysis in chemometrics, CRC Press, 2009,
Chapter 2.4 Distances and Similarities.

Theodoridis, S. and Koutroumbas, K.: Pattern Recognition Academic Press, 2006,
Chapter 11.2.2 Proximity Measures between Two Points.
}

