% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{elink}
\alias{elink}
\title{Exact linkage algorithm}
\usage{
elink(ss)
}
\arguments{
\item{ss}{A numeric matrix with the assignments for each individual (rows)
in each iteration (columns), as output by \code{DPM_sampler}}
}
\value{
A list with two elements: a two-column matrix with pairs of
individuals (one per row) and a numeric vector with the distances between
each pair
}
\description{
This function uses an exact linkage algorithm to compute the distance
between pairs of individuals, based on the set of class assignments produced
by \code{DPM_sampler}.
}
