% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\name{gaussian_model}
\alias{gaussian_model}
\title{Constructor for Gaussian models}
\usage{
gaussian_model(n_components, type = "gaussian_unit")
}
\arguments{
\item{n_components}{Integer. The number of latent classes/components to estimate.}

\item{type}{Character. The specific variance structure, e.g., "gaussian_diag" or "gaussian_unit".}
}
\value{
A list object containing the model state.
}
\description{
Sets up the initial state and class structure for continuous emission models
(like Gaussian with diagonal or unit variance).
}
