% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLD.r
\name{KLD}
\alias{KLD}
\title{Kullback-Leibler divergence between priors}
\usage{
KLD(ab, params)
}
\arguments{
\item{ab}{A vector with the initial values for the Gamma(a, b) prior
parameters}

\item{params}{A list of fixed parameters. See details.}
}
\value{
An atomic numeric vector with the value of the Kullback-Leibler
divergence
}
\description{
Calculates the Kullback-Leibler divergence between a probability mass
function \eqn{g(k)} and a \eqn{Gamma(a, b)} prior for the alpha parameter.
}
\details{
This function is not meant to be used by itself, but instead to be called by
the \code{gamma_prior} function, which tries to find the values for
\eqn{a} and \eqn{b} that minimize the Kullback-Leibler divergence for a
given \eqn{k} and \eqn{n}.

Besides \code{k} and \code{n}, a matrix of unsigned Stirling numbers of the
first kind for \eqn{n} elements and \eqn{n} cycles must also be passed as a
fixed parameter.
}
\references{
Dorazio, R.M. (2009). On selecting a prior for the precision parameter of
Dirichelet process mixture models. \emph{Journal of Statistical Planning and
Inference}, \strong{139}, 3384-3390. doi:10.1016/j.jspi.2009.03.009
}
\seealso{
\link{gamma_prior}, \link{stirling1}
}
