\name{nfindrCB}
\alias{nfindrCB}
\title{Fast N-FINDR Variant}
\usage{
  nfindrCB(data, p, indices, ..., iters = 3)
}
\arguments{
  \item{data}{Data matrix to unmix}

  \item{p}{Number of endmembers}

  \item{indices}{Array of indices generated by
  \code{\link{nfindr}}. Its contents are ignored but it is
  used to store the real endmember indices}

  \item{...}{Extra unused parameters that get passed in
  from \code{\link{nfindr}}}

  \item{iters}{Max number of iterations, defaults to 3*p}
}
\value{
  The indices that indicate the position of the endmembers
  in the original dataset
}
\description{
  (DESCRIPTION)
}

