% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{nfindrSeqLDU}
\alias{nfindrSeqLDU}
\title{LDU-Sequential N-FINDR: Slightly modified LDU N-FINDR}
\usage{
nfindrSeqLDU(data, p, indices, ...)
}
\arguments{
\item{data}{Data matrix to unmix}

\item{p}{Number of endmembers}

\item{indices}{Indices used in the simplex estimation}

\item{...}{Extra unused parameters that get passed in from
\code{\link{nfindr}}}
}
\value{
The indices that indicate the position of the endmembers in the
  original dataset
}
\description{
While LDU N-FINDR examines a single pixel in each endmember position and
repeats over all pixels, this algorithm considers all pixels in a single
endmember position and then repeats over all endmember positions.
Intended to be called from \code{\link{nfindr}}.
}
\references{
Dowler, Shaun W.; Takashima, Reymond; Andrews, Mark;
  "Reducing the complexity of the N-FINDR algorithm for hyperspectral
  image analysis.", IEEE Trans Image Process. 2013 Jul;22(7):2835-48l
  doi: 10.1109/TIP.2012.2219546
}

