% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/options.R
\name{.options}
\alias{.options}
\alias{unmixR.options}
\title{unmixR's package options}
\usage{
.options(..., .__defaults = FALSE, .__reset = FALSE)

unmixR.options(...)
}
\arguments{
\item{...}{either \code{key = value} pairs to set options or the names of
the options to retrieve. If no paramaters are passed, a list of all
options is returned.}
}
\value{
list of options
}
\description{
\pkg{unmixR} uses \pkg{settings} for option management. Options can either
be set by \code{unmixR.options (key = value)}
}
\details{
The following package-specific options are defined:

\tabular{lll}{
debuglevel \tab >= 0L \tab indicates how much debuging output is to be produced.
}
}
\examples{
unmixR.options ()
unmixR.options ("debuglevel")
unmixR.options (debuglevel = 0L)
}

