% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\alias{predict.nfindr}
\alias{predict.vca}
\title{Predict Endmember Abundances}
\usage{
\method{predict}{nfindr}(object, newdata = object$data, ...)

\method{predict}{vca}(object, newdata = object$data, ...)
}
\arguments{
\item{object}{The N-FINDR/VCA structure returned by the
\code{\link{nfindr}}/\code{\link{vca}} interface}

\item{newdata}{If the data stored in the object is not the data that
should be checked for abundances then this parameter allows for passing
in new data}

\item{...}{Allow for extra parameters to match the signature of the base
predict function}
}
\value{
A matrix where the abundances for an endmember are returned
  column-wise. Each value is in the range \code{0 - 1}
}
\description{
Predicts the abundance percentages of each endmember at all sample points
using the Non-Negative Least Squares method
}

