\docType{package}
\name{spec.unmixing-package}
\alias{spec.unmixing-package}
\title{Spectral Unmixing Methods}
\description{
  Implements the N-FINDR and Vertex Component Analysis
  algorithms to recover pure component spectra and their
  respective concentrations from a set of measured spectra.
}
\author{
  Conor McManus

  Maintainer: Conor McManus <chathurga@gmail.com>
}
\keyword{package}

