% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stirling1.r
\name{stirling1}
\alias{stirling1}
\title{Stirling numbers of the first kind}
\usage{
stirling1(n, k)
}
\arguments{
\item{n}{Number of elements.}

\item{k}{Number of orbits.}
}
\value{
A lower triangular matrix of Stirling numbers of the first kind for
every pair of (n, k).
}
\description{
Calculation of (signed) Stirling numbers of the first kind for a given number
of elements and orbits.
}
\examples{
stirling1(4, 5)

}
