% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{package}
\name{unmixR-package}
\alias{unmixR}
\alias{unmixR-package}
\title{Hyperspectral Unmixing Methods}
\description{
unmixR implements the N-FINDR and Vertex Component Analysis (VCA)
algorithms which can recover pure component spectra and their
respective concentrations from a hyperspectral data set.
}
\keyword{package}

