% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unittests.R
\name{unmixR.unittest}
\alias{unmixR.unittest}
\title{Run the unit tests}
\usage{
unmixR.unittest()
}
\value{
invisibly \code{TRUE} if the tests pass, \code{NA} if \link[svUnit]{svUnit} is not
available. Stops if errors are encountered.
}
\description{
Run the unit tests attached to the functions via \link[svUnit]{svUnit}
}
\author{
Claudia Beleites
}
\seealso{
\link[svUnit]{svUnit}
}
\keyword{programming}
\keyword{utilities}

