% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vcaLopez.R
\name{vcaLopez}
\alias{vcaLopez}
\title{Lopez VCA}
\usage{
vcaLopez(data, p)
}
\arguments{
\item{data}{Data to unmix. It will be converted to a matrix using
as.matrix. The matrix should contain a spectrum per row.}

\item{p}{Number of endmembers}
}
\value{
Endmember indices
}
\description{
The original VCA algorithm (\code{\link{vca}}) contains auxiliary
operations (signal to noise estimation, etc.) which could be omitted while
still maintaining a fully functional algorithm. In paper that describes
the Modified VCA algorithm (\code{\link{vcaMod}}) the author provides a
description of the original VCA with only the essential components
included. This is an implementation of that algorithm
}
\references{
Lopez, S.; Horstrand, P.; Callico, G.M.; Lopez, J.F.;
Sarmiento, R., "A Low-Computational-Complexity Algorithm for
Hyperspectral Endmember Extraction: Modified Vertex Component Analysis,"
Geoscience & Remote Sensing Letters, IEEE, vol.9 no.3 pp.502,506, May 2012
doi: 10.1109/LGRS.2011.2172771
}

