% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{elink}
\alias{elink}
\title{Exact linkage algorithm}
\usage{
elink(x)
}
\arguments{
\item{x}{A matrix with the class assignments for each individual (rows) in '
each iteration (rows)}
}
\value{
a \code{list} containing two objects: \code{pairs}, a matrix with
the pairs of leaves/groups that make up the tree, and \code{distances}, a
numeric vector with the distance between the elements of each pair
}
\description{
Algorithm by Dawnson & Belkir (2009) to construct a tree based on estimates
of marginal co-assignment probabilities.
}
