% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_prior.r
\name{gamma_prior}
\alias{gamma_prior}
\title{Prior parameters for gamma function}
\usage{
gamma_prior(n, distr = c("unif", "pois", "nbinom", "lnorm", "norm"), ...)
}
\arguments{
\item{n}{Sample size}

\item{distr}{The probability distribution for the prior}

\item{...}{other parameters passed to the distribution function}
}
\value{
A names vector with parameters \code{a} and \code{b}
}
\description{
Prior parameters for gamma function
}
\examples{
gamma_prior(150, distr = 'pois', lambda = 3)
}
