\name{hfc}
\alias{hfc}
\title{Harsanyi-Farrand-Chang method for estimating the number of endmembers in
a spectral image}
\usage{
  hfc(data, alpha)
}
\arguments{
  \item{data}{Spectra matrix to examine}

  \item{alpha}{False-alarm probability}
}
\value{
  (-return-)
}
\description{
  (-description-)
}

