% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{spectralDist}
\alias{spectralDist}
\title{Compute distances between spectra}
\usage{
spectralDist(M = stop("Missing matrix of spectra"), method = c("euclidean",
  "maximum", "manhattan", "canberra", "binary", "minkowski", "pearson",
  "correlation", "spearman", "kendall", "abspearson", "abscorrelation",
  "cosineAlpha"), ref, reportAsRank = FALSE, ...)
}
\arguments{
\item{M}{Matrix containing spectra to be compared in rows.
The matrix should have dimensions no. of samples x
no. of wavelengths.}

\item{method}{String.  The method to be used in computing the distance.
One of \code{c("name them")}.}

\item{ref}{Defaults to \code{NULL}.  Otherwise, an integer
giving the row of M to be used as the reference spectrum.}

\item{reportAsRank}{If \code{ref} is given, this flag will
cause the vector to be reported as a rank.}

\item{...}{Arguments to be passed downstream, for instance, the power
for the Minkowski computation.  See \code{\link{dist}} for details.}
}
\value{
pair-wise distance matrix of all cases in \code{M} against those indicated by \code{ref}.

If \code{ref} is not given, a dist object with all pair-wise distances.

If \code{reportAsRank = TRUE} the vector will be reported as a ranked vector.
}
\description{
Compute the distances between spectra in
a set, providing a measure of their similarity.
Common distance measures are available.
Optionally, a particular spectrum may be specified as
a reference spectrum.  In all cases it is assumed
that the spectra are in similar units and appropriately
scaled.
}
\author{
Bryan Hanson, with contributions by Claudia Beleites
}
\seealso{
\code{\link[amap]{Dist}}, \code{\link[stats]{dist}},  \code{\link{cos.dist}}
}

