% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unmixR-package.R
\docType{package}
\name{unmixR-package}
\alias{unmixR-package}
\title{Spectral Unmixing Methods}
\description{
Implements the N-FINDR and Vertex Component Analysis algorithms
to recover pure component spectra and their respective concentrations
from a set of measured spectra.
}
\author{
Conor McManus

Maintainer: Conor McManus <chathurga@gmail.com>
}
\keyword{package}

