% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vca05.R
\name{vca05}
\alias{vca05}
\title{Vertex Component Analysis unmixing algorithm}
\usage{
vca05(data, p, SNR = estSNR(data, p))
}
\arguments{
\item{data}{Data to unmix. It will be converted to a matrix using
as.matrix. The matrix should contain a spectrum per row.}

\item{p}{Number of endmembers}

\item{SNR}{The Signal-to-Noise ratio of the data. By default it will be
  estimated using \code{\link{estSNR}}}
}
\value{
Although the other VCA algorithms return only indices, this
  function returns the full structure to \link{\code{vca}} as the
  endmembers should be taken out of a projected dataset which it generates
}
\description{
This algorithm is based on the geometry of convex sets. It exploits the
fact that endmembers occupy the vertices of a simplex.
}
\references{
Nascimento, J.M.P.; Bioucas Dias, J.M., "Vertex component
  analysis: a fast algorithm to unmix hyperspectral data," Geoscience and
  Remote Sensing, vol.43, no.4, pp.898,910, April 2005;
  doi: 10.1109/TGRS.2005.844293
}

